Switched css framework to spectre.css + added tooling
The css framework switch meant changing all old files to use the new classes.
In order to override the colors in spectre.less, custom styles will be written in less from now on. Compiling less in now another step in making the app, so the README.md was updated to reflect that.