Major refactoring of playback commands
Other commands should work the same as play and pause currently do. Still need to figure out the best way to get the response after a command has been called. It's going to require checking the last command sent and knowing that the next message to come back from the websocket is going to be the response of the command.